fix(hw16): use an alternative location for postgresql.conf
Otherwise we get "initdb: error: directory "/var/lib/postgresql/data"
exists but is not empty" on a clean postgresql container restart (it
sees the file postgresql.conf that is already there).

https://stackoverflow.com/a/70079671

The solution is to mount the external postgresql.conf to a different
path in the container and communicate it to postgres server.

https://stackoverflow.com/a/41912295