feat(hw14): set up monitoring stack
Includes Prometheus, Node Exporter and Grafana running in Docker
containers. Prometheus is already configured as a datasource in Grafana.
In turn Prometheus has Node Exporter instance in it's scraping config.
This way it's pretty easy to quickly set up a dashboard for host system
monitoring in Grafana (e.g. 11074).

I was generally following an example from this article:
https://last9.io/blog/prometheus-with-docker-compose/