chore(hw14): recover full URL in links to the current repo
While resources (images, hyperlinks) with abbreviated URLs (with
github.com domain name omitted) are accessed correctly from Wiki
markdown (the images are rendered etc.) they are not available if the
corresponding .md file is opened in the repository code tree (the
eventual URL is not formed correctly). The easiest way is just to use
full URL with github.com in the first place.